rss_aggregator_function/requirements.txt
requests
feedparser
SQLAlchemy
psycopg2-binary
google-cloud-pubsub
google-generativeai